Split out test_event_handlers
